Moved InterfaceInstance and InterfaceBase types into different files
