+++ src/ast/stmt/record.rs
@@ -2,7 +2,7 @@ use crate::{
     ast::{arg::UntypedArg, member::UntypedMember},
     error::HayError,
     lex::token::Token,
-    types::{base::validate_requirements, RecordKind, Type, TypeId, TypeMap},
+    types::{validate_requirements, RecordKind, Type, TypeId, TypeMap},
 };
 
 use super::Stmt;
