+++ src/ast/stmt/stub.rs
@@ -2,7 +2,7 @@ use crate::{
     ast::arg::UntypedArg,
     error::HayError,
     lex::token::Token,
-    types::{base::validate_requirements, FnTag, FunctionStub, Signature, Type, TypeId, TypeMap},
+    types::{validate_requirements, FnTag, FunctionStub, Signature, Type, TypeId, TypeMap},
 };
 
 use super::{GlobalEnv, Stmt, StmtKind};
