+++ src/types/interface/mod.rs
@@ -1,3 +1,7 @@
-pub mod associated_type;
-pub mod base;
-pub mod instance;
+mod associated_type;
+mod base;
+mod instance;
+
+pub use associated_type::*;
+pub use base::*;
+pub use instance::*;
